sycl : fix compilation warnings and remove unused variables
- Add [[maybe_unused]] to unused parameters in kernel functions
- Fix lambda captures to not capture global variables unnecessarily
- Use static_cast instead of C-style casts for const correctness
- Fix pointer-bool-conversion warnings (tensor->name[0] instead of tensor->name)
- Remove unused static variable g_tp_dev1_kv_cache_initialized
- Add explicit float cast in debug output (mmvq.cpp)

Tested with Mistral 7B Q4_0 and GPT-OSS 20B Q8_0 models.

+++ ggml/src/ggml-sycl/mmvq.cpp
@@ -1375,7 +1375,7 @@ bool ggml_sycl_mul_mat_id_vec_q(
                     for (int i = 0; i < 16; i++) {
                         fprintf(stderr, "%02x ", host_mxfp4.qs[i]);
                     }
-                    fprintf(stderr, "\n[MXFP4 DEBUG] Q8_1 d=%f, qs: ", host_q8_1.ds[0]);
+                    fprintf(stderr, "\n[MXFP4 DEBUG] Q8_1 d=%f, qs: ", static_cast<float>(host_q8_1.ds[0]));
                     for (int i = 0; i < 32; i++) {
                         fprintf(stderr, "%d ", host_q8_1.qs[i]);
                     }
